producer/requirements.txt
newsapi-python
kafka-python